% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_cols.r
\name{combine_cols}
\alias{combine_cols}
\title{Combine Factors Correctly}
\usage{
combine_cols(.x, .y, .keep.factor = TRUE)
}
\arguments{
\item{.x}{A vector}

\item{.y}{A vector}

\item{.keep.factor}{Nothing yet}
}
\description{
Just c(.x, .y) will produce incorrect results of .x or .y is a factor. This
is a replacement that handles factors correctly (hopefully).
}

