% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dt_state.r
\name{get_dt_state}
\alias{get_dt_state}
\alias{set_dt_state}
\title{Save and Restore the Current State of a Data Structure}
\usage{
get_dt_state(.data)

set_dt_state(.data, .state)
}
\arguments{
\item{.data}{The data structure}

\item{.state}{The return value of an earlier get_dt_state call}
}
\description{
This is much less grandiose than the title. It just saves whether you have
a data.table, data.frame or list and if it's a data.table also saves the
key columns.
}
\examples{
library(data.table)
DT <- data.table(x=1:10, key="x")

state <- BigStuff:::get_dt_state(DT)
BigStuff:::set_dt_state(DT, state)
}

