% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setList.r
\name{setList}
\alias{setList}
\title{Turn a data.table or data.frame into a list by Reference}
\usage{
setList(.data)
}
\arguments{
\item{.data}{A data.table or data.frame}
}
\value{
This will change the type by reference. Nevertheless it is also
returned invisibly for ease of use with magrittr and to be consistent with
setDT and setDF.
}
\description{
This is analogous to setDT and setDF from data.table, although far less
elaborate.
}
\examples{
DF <- data.frame(x=1:10)
setList(DF)
stopifnot(is.data.frame(DF) == FALSE)
}

